requirements.txt
backcall==0.1.0
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
decorator==4.4.1
Django==2.1.1
django-bootstrap4==1.0.1
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.10
Pygments==2.4.2
python-decouple==3.3
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
